Bring back a lost commit
When we switched developer during vacation period, a whole commit was left
behind. That resulted in bnc#946639, some test coverage decrease and another
subtle bug (not reported yet).

This commits brings back the missing bits from that lost commit
(the rest was already reimplemented as a fix for bnc#946639).

The commit also contained a quite lengthly message, reproduced below:

"It actually has been a long time since the code for removing offsets was
 included in any workflow (i.e. used at all). This commit not only fixes the
 code (broken by the latests changes), but also ensures it's called again"