Do not let users change the allocation with auto resize
The whole point of KDUMP_AUTO_RESIZE is that users do not have to
set any values. Let YaST calculate a sensible allocation
automatically.

The current logic is to take the size of the largest available
memory block, or half of total RAM, whichever is smaller.

For Xen Dom0, we take the default values, because the Xen
hypervisor cannot shrink the crashkernel area after boot.